refactor: move `RpcConfig` to miden-cli
