+++ tests/integration/main.rs
@@ -799,8 +799,8 @@ async fn test_get_account_update() {
     // Request updates from node for both accounts. The request should not fail and both types of
     // [AccountDetails] should be received.
     // TODO: should we expose the `get_account_update` endpoint from the Client?
-    let (rpc_config, _) = get_client_config();
-    let mut rpc_api = TonicRpcClient::new(&rpc_config);
+    let (endpoint, timeout, _) = get_client_config();
+    let mut rpc_api = TonicRpcClient::new(endpoint, timeout);
     let details1 = rpc_api.get_account_update(basic_wallet_1.id()).await.unwrap();
     let details2 = rpc_api.get_account_update(basic_wallet_2.id()).await.unwrap();
